Refactored fertiliser model
